src/third_party/blink/web_tests/external/wpt/css/css-text-decor/parsing/text-decoration-computed-expected.txt

This is a testharness.js-based test.
FAIL Property text-decoration value 'none' assert_equals: expected "rgb(0, 0, 255)" but got "none solid rgb(0, 0, 255)"
FAIL Property text-decoration value 'line-through' assert_equals: expected "line-through rgb(0, 0, 255)" but got "line-through solid rgb(0, 0, 255)"
FAIL Property text-decoration value 'solid' assert_equals: expected "rgb(0, 0, 255)" but got "none solid rgb(0, 0, 255)"
FAIL Property text-decoration value 'currentcolor' assert_equals: expected "rgb(0, 0, 255)" but got "none solid rgb(0, 0, 255)"
PASS Property text-decoration value 'double overline underline'
FAIL Property text-decoration value 'underline overline line-through red' assert_equals: expected "underline overline line-through rgb(255, 0, 0)" but got "underline overline line-through solid rgb(255, 0, 0)"
FAIL Property text-decoration value 'rgba(10, 20, 30, 0.4) dotted' assert_equals: expected "dotted rgba(10, 20, 30, 0.4)" but got "none dotted rgba(10, 20, 30, 0.4)"
PASS Property text-decoration value 'underline dashed rgb(0, 255, 0)'
FAIL Property text-decoration value 'auto' assert_true: 'auto' is a supported value for text-decoration. expected true got false
FAIL Property text-decoration value 'from-font' assert_true: 'from-font' is a supported value for text-decoration. expected true got false
FAIL Property text-decoration value '10px' assert_true: '10px' is a supported value for text-decoration. expected true got false
FAIL Property text-decoration value 'underline red from-font' assert_true: 'underline red from-font' is a supported value for text-decoration. expected true got false
Harness: the test ran to completion.
